+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 - Added initiator data (manual or automated) to `site backups list`. (#716)
 - New command `site backups get-schedule` shows the scheduled weekly backup day and daily backup time. (#723)
 - New command `workflows watch` to stream workflow updates (#722)
+- New command `site backups set-schedule` schedules the daily backup and weekly day. (#724)
 
 ### Fixed
 - `site backups get` no longer errs when there are no backups. (#690)